+++ tests/arp/arp_utils.py
@@ -34,26 +34,6 @@ def collect_info(duthost):
         duthost.shell('grep . /sys/class/net/PortChannel*/address', module_ignore_errors=True)
 
 
-def increment_ipv4_addr(ipv4_addr, incr=1):
-    octets = str(ipv4_addr).split('.')
-    last_octet = int(octets[-1])
-    last_octet += incr
-    octets[-1] = str(last_octet)
-
-    return '.'.join(octets)
-
-
-def increment_ipv6_addr(ipv6_addr, incr=1):
-    octets = str(ipv6_addr).split(':')
-    last_octet = octets[-1]
-    if last_octet == '':
-        last_octet = '0'
-    incremented_octet = int(last_octet, 16) + incr
-    new_octet_str = '{:x}'.format(incremented_octet)
-
-    return ':'.join(octets[:-1]) + ':' + new_octet_str
-
-
 def MacToInt(mac):
     mac = mac.replace(":", "")
     return int(mac, 16)
