+++ tests/generic_config_updater/test_mgmt_interface.py
@@ -3,24 +3,16 @@ import logging
 import pytest
 
 from tests.common.helpers.assertions import pytest_assert
-from tests.common.utilities import wait_until
 from tests.generic_config_updater.gu_utils import apply_patch, create_path
 from tests.generic_config_updater.gu_utils import generate_tmpfile, delete_tmpfile
 from tests.generic_config_updater.gu_utils import create_checkpoint, delete_checkpoint, rollback_or_reload
+from tests.common.utilities import wait_for_file_changed, FORCED_MGMT_ROUTE_PRIORITY
 
 pytestmark = [
     pytest.mark.topology('any'),
 ]
 
 logger = logging.getLogger(__name__)
-
-# forced mgmt route priority hardcoded to 32764 in following j2 template:
-# https://github.com/sonic-net/sonic-buildimage/blob/master/files/image_config/interfaces/interfaces.j2#L82
-FORCED_MGMT_ROUTE_PRIORITY = 32764
-
-# Wait 300 seconds because sometime 'interfaces-config' service take 45 seconds to response
-# interfaces-config service issue track by: https://github.com/sonic-net/sonic-buildimage/issues/19045
-FILE_CHANGE_TIMEOUT = 300
 
 
 @pytest.fixture(scope="function")
@@ -40,36 +32,6 @@ def ensure_dut_readiness(duthost):
         rollback_or_reload(duthost)
     finally:
         delete_checkpoint(duthost)
-
-
-def get_file_hash(duthost, file):
-    hash = duthost.command("sha1sum {}".format(file))["stdout"]
-    logger.debug("file hash: {}".format(hash))
-
-    return hash
-
-
-def get_interface_reload_timestamp(duthost):
-    timestamp = duthost.command("sudo systemctl show --no-pager interfaces-config"
-                                " -p ExecMainExitTimestamp --value")["stdout"]
-    logger.info("interfaces config timestamp {}".format(timestamp))
-
-    return timestamp
-
-
-def wait_for_file_changed(duthost, file, action, *args, **kwargs):
-    original_hash = get_file_hash(duthost, file)
-    last_timestamp = get_interface_reload_timestamp(duthost)
-
-    action(*args, **kwargs)
-
-    def hash_and_timestamp_changed(duthost, file):
-        latest_hash = get_file_hash(duthost, file)
-        latest_timestamp = get_interface_reload_timestamp(duthost)
-        return latest_hash != original_hash and latest_timestamp != last_timestamp
-
-    exist = wait_until(FILE_CHANGE_TIMEOUT, 1, 0, hash_and_timestamp_changed, duthost, file)
-    pytest_assert(exist, "File {} does not change after {} seconds.".format(file, FILE_CHANGE_TIMEOUT))
 
 
 def update_forced_mgmt_route(duthost, interface_address, interface_key, routes):
