+++ tests/route/test_forced_mgmt_route.py
@@ -5,7 +5,8 @@ import pytest
 
 from tests.common.config_reload import config_reload
 from tests.common.helpers.assertions import pytest_assert
-from tests.common.utilities import wait_until
+from tests.common.utilities import wait_until, wait_for_file_changed, \
+    FORCED_MGMT_ROUTE_PRIORITY
 from tests.override_config_table.utilities import backup_config, restore_config, \
         reload_minigraph_with_golden_config
 from tests.syslog.syslog_utils import is_mgmt_vrf_enabled
@@ -17,16 +18,6 @@ pytestmark = [
 ]
 
 logger = logging.getLogger(__name__)
-
-
-# forced mgmt route priority hardcoded to 32764 in following j2 template:
-# https://github.com/sonic-net/sonic-buildimage/blob/master/files/image_config/interfaces/interfaces.j2#L82
-FORCED_MGMT_ROUTE_PRIORITY = 32764
-
-
-# Wait 300 seconds because sometime 'interfaces-config' service take 45 seconds to response
-# interfaces-config service issue track by: https://github.com/sonic-net/sonic-buildimage/issues/19045
-FILE_CHANGE_TIMEOUT = 300
 
 
 @pytest.fixture
@@ -44,36 +35,6 @@ def backup_restore_config(duthosts, enum_rand_one_per_hwsku_hostname):
     #  Restore config after test finish
     restore_config(duthost, CONFIG_DB, CONFIG_DB_BACKUP)
     config_reload(duthost)
-
-
-def get_interface_reload_timestamp(duthost):
-    timestamp = duthost.command("sudo systemctl show --no-pager interfaces-config"
-                                " -p ExecMainExitTimestamp --value")["stdout"]
-    logger.info("interfaces config timestamp {}".format(timestamp))
-
-    return timestamp
-
-
-def get_file_hash(duthost, file):
-    hash = duthost.command("sha1sum {}".format(file))["stdout"]
-    logger.debug("file hash: {}".format(hash))
-
-    return hash
-
-
-def wait_for_file_changed(duthost, file, action, *args, **kwargs):
-    original_hash = get_file_hash(duthost, file)
-    last_timestamp = get_interface_reload_timestamp(duthost)
-
-    action(*args, **kwargs)
-
-    def hash_and_timestamp_changed(duthost, file):
-        latest_hash = get_file_hash(duthost, file)
-        latest_timestamp = get_interface_reload_timestamp(duthost)
-        return latest_hash != original_hash and latest_timestamp != last_timestamp
-
-    exist = wait_until(FILE_CHANGE_TIMEOUT, 1, 0, hash_and_timestamp_changed, duthost, file)
-    pytest_assert(exist, "File {} does not change after {} seconds.".format(file, FILE_CHANGE_TIMEOUT))
 
 
 def address_type(address):
